+++ client/src/scripts/others/config.ts
@@ -78,6 +78,7 @@ export const ConfigCasters=Object.freeze({
     sv_ui_translation:Casters.toString,
     sv_ui_interactive:Casters.toBoolean,
     sv_ui_blur_backdrop:Casters.toBoolean,
+    sv_ui_simple_mode:Casters.toBoolean,
 
     sv_debug_ping_emulation:Casters.toNumber,
 })
@@ -100,7 +101,7 @@ export const ConfigDefaultValues={
     sv_graphics_resolution:(Debug.force_mobile||isMobile)?"low":"medium",
     sv_graphics_particles:GraphicsDConfig.Advanced,
     sv_graphics_lights:GraphicsDConfig.Advanced,
-    sv_graphics_post_proccess:GraphicsDConfig.Advanced,
+    sv_graphics_post_proccess:(Debug.force_mobile||isMobile)?GraphicsDConfig.None:GraphicsDConfig.Advanced,
     sv_graphics_climate:true,
     sv_graphics_fullscreen:false,
 
@@ -111,7 +112,7 @@ export const ConfigDefaultValues={
     sv_game_ammo_outline:false,
     sv_game_ping:5,
 
-    sv_mobile_auto_pickup:(Debug.force_mobile||isMobile),
+    sv_mobile_auto_pickup:Debug.force_mobile||isMobile,
 
     sv_sounds_master_volume:1,
     sv_sounds_music_volume:1,
@@ -126,7 +127,8 @@ export const ConfigDefaultValues={
     sv_ui_special_color:"#fffb00",
     sv_ui_translation:"en",
     sv_ui_interactive:true,
-    sv_ui_blur_backdrop:true,
+    sv_ui_blur_backdrop:!(Debug.force_mobile||isMobile),
+    sv_ui_simple_mode:Debug.force_mobile||isMobile,
 
     sv_debug_ping_emulation:0,
 }