+++ common/scripts/config/final_screen.ts
@@ -32,46 +32,6 @@ export interface FinalScreenDef{
     foreground?:FinalScreenLayer[]
     background?:FinalScreenLayer[]
 }
-export const city_final:FinalScreenDef={
-    theme:{
-        css:`
-            background: linear-gradient(to bottom, #0d1e35, #040618);
-        `,
-        accent:"#33b8ff",
-        music:"/sounds/musics/final_screen_music_1.mp3"
-    },
-    background:[
-        {
-            type:FinalScreenLayerType.Static,
-            image:"/img/menu/final_screen/moon.svg",
-            css:`
-                top:150px;
-                left:50%;
-                transform: translateX(-50%)
-            `
-        },
-        {
-            type:FinalScreenLayerType.Walk,
-            image:"/img/menu/final_screen/building_1.svg",
-            css:`
-                top:280px;
-            `,
-            speed:0.7,
-            count:2,
-            gap:1500,
-        },
-        {
-            type:FinalScreenLayerType.Walk,
-            image:"/img/menu/final_screen/palm.svg",
-            css:`
-                top:600px;
-            `,
-            speed:1,
-            count:4,
-            gap:800,
-        }
-    ],
-}
 export const island_final:FinalScreenDef={
     theme:{
         css:`
@@ -83,22 +43,23 @@ export const island_final:FinalScreenDef={
     background:[
         {
             type:FinalScreenLayerType.Static,
-            image:"/img/menu/final_screen/moon.svg",
+            image:"/img/menu/final_screen/moon.png",
             css:`
                 top:150px;
                 left:50%;
+                height:200px;
                 transform: translateX(-50%);
             `
         },
-        // Reflected Moon
         {
             type:FinalScreenLayerType.Static,
-            image:"/img/menu/final_screen/moon.svg",
+            image:"/img/menu/final_screen/moon.png",
             css:`
+                top:150px;
                 left:50%;
-                bottom:0px;
-                transform: translateX(-50%) scale(65%, -65%);
-                filter: blur(20px);
+                height:200px;
+                transform: translateX(-50%);
+                filter: blur(100px);
             `
         },
         // Cloud
@@ -114,38 +75,59 @@ export const island_final:FinalScreenDef={
             count:3,
             gap:1000,
         },
-        {
-            type:FinalScreenLayerType.Walk,
-            image:"/img/menu/final_screen/cloud_1.svg",
-            css:`
-                bottom:25px;
-                opacity:50%;
-                filter: blur(10px);
-                transform: scale(60%,-60%);
-            `,
-            speed:0.01,
-            count:3,
-            gap:1000,
-        },
         // Mountains
         {
             type:FinalScreenLayerType.Tile,
-            image:"/img/menu/final_screen/mountains_1.svg",
+            image:"/img/menu/final_screen/mountains_1.png",
             css:`
-                bottom:250px;
-                height:400px;
+                bottom:0px;
+                height:750px;
             `,
-            speed:0.05
+            speed:0.04
         },
-        
+
+        {
+            type:FinalScreenLayerType.Walk,
+            image:"/img/menu/final_screen/palm.png",
+            css:`
+                height:320px;
+                bottom:-20px;
+            `,
+            speed:0.35,
+            count:4,
+            gap:800,
+        },
+        {
+            type:FinalScreenLayerType.Walk,
+            image:"/img/menu/final_screen/building_1.png",
+            css:`
+                height:550px;
+                bottom:0px;
+            `,
+            speed:0.5,
+            count:5,
+            gap:1000,
+        },
+        {
+            type:FinalScreenLayerType.Walk,
+            image:"/img/menu/final_screen/palm.png",
+            css:`
+                height:480px;
+                bottom:-20px;
+            `,
+            speed:0.75,
+            count:4,
+            gap:800,
+        }
+        /*
         // Reflected Montains
         {
             type:FinalScreenLayerType.Tile,
-            image:"/img/menu/final_screen/mountains_1.svg",
+            image:"/img/menu/final_screen/mountains_1.png",
             css:`
                 bottom:250px;
                 height:400px;
-                transform: scaleY(-50%) translateY(-100%);
+                transform: scale(100%,-50%), translateY(-100%);
                 filter: blur(20px);
             `,
             speed:0.05
@@ -184,17 +166,7 @@ export const island_final:FinalScreenDef={
             `,
             speed:0.11
         },
-        {
-            type:FinalScreenLayerType.Tile,
-            image:"/img/menu/final_screen/waves.svg",
-            css:`
-                bottom:295px;
-                height:30px;
-                opacity:50%;
-                filter: blur(1px);
-            `,
-            speed:0.1
-        },
+        */
     ],
     foreground:[
     ],